Internationalization (i18n) support. Currently works only with Hungarian language but any language can be easily added. See src/i18n/README for details.
The "configure" script has been also updated. With autoconf 2.52 I also
regenerated this file. I also used autoscan to generate an up-to-date
configure.in and I merged the old one together with the new one.
Hopefully it still works now without any problems.

Bugs: if gettext is not installed, the package does not work anymore.
There should be checked if it is missing or not, mainly under DOS and
Windows (and maybe other platforms as well) and if it is not available,
then easily add a

STRING gettext(STRING translatable)
{
 return translatable;
}

or something like this to the code with #ifdef.