Here I attach the whole information text about this patch. This improvement was written by Paul Nasca <paulnasca@email.ro>. Note that an internationalized string should be also changed.
Zoltan



Xaos version: Xaos-3.0

*** NOTE (Zoltan) ***
This is not the newest version.
**** END OF NOTE ****

the file:  src/ui/uinew.c

*** NOTE (Zoltan) ***
This file no longer exists, there is ui.c instead of it.
**** END OF NOTE ****

This patch does 2 improvements:
     1) the number of iterations are no longer incremented/decremendet by one. This is very annoing if I want to change from 10 to 100,000.
        Instead it is inc./dec. by about 5 percent of the current values.
	So, I can change in a better way the number of iterations.
	The problem is that the you cannot change it very preciselly at high numbers (anyway who cares if the number of iterations are 100,000 or 100,001 ?)

*** NOTE (Zoltan) ***
In version 3.1 this is an obsolete improvement since it already contains such a thing.
**** END OF NOTE ****

     2) the status line show the logarithm of the zoom value like:
        (zoomed 1281.25 times (3.1E) ....)
	It is very usefull at high zoomings:
	Eg. Instead counting the digit of the number 3447484602761.55 better you read (12.5E) and you know that is zoomed about 10**12 times.

I'd like if you'll add this in the xaos distribution. Hope you'll enjoy it.

Paul Nasca from Romania.

P.S. Thank you for this powerful program. When I saw first time on a 486 I could'nt belive that is possible to do a real-time zooming of fractals (even on very poverfull computers).