There are 3 new fractal types and an option to change the bailout value. The improvements have been made by Andreas Madritsch.
All modified files were autoindented on Red Hat 6.1,
however all other files were autoindented on SuSE 7.3.
To be consistent, these files have to be reindented
on SuSE 7.3.

Here I also attach the original mail written by
Andreas.

Zoltan



From amadritsch@datacomm.ch Fri Sep 27 15:27:45 2002
Date: Fri, 27 Sep 2002 01:40:09 +0200
From: Andreas Madritsch <amadritsch@datacomm.ch>
To: kovzol@math.u-szeged.hu
Cc: jh@suse.cz
Subject: new features to XaoS

Hello

Here are my new features.

First, I added three new fractal types named newton^4, barnsley2 and
magnet2 to XaoS. For this, I made changes to the file engine/formulas.c.
I have added the original file formulas.c.orig and the changed version
formulas.c.1 to the tar archive attached, so that you can compare them
with a compare program and point out my changes. To avoid conflicts with
existing shortkeys, I changed the file ui-hlp/menu.c that the shortkeys
for additional fractal types become A, B, C in capital letters and so
on. As far as I know, capital letters are not used by XaoS yet. Because
you would use option-A, option-B and option-C on a Mac, which is not case
sensitive and conflicts with autopilot, perturbation and the outcoloring
menu, I switched them of for Mac, so Mac users have to use the menu.
There are also the files menu.c.orig and menu.c.1 in the tar archive, to
examine the changes I made.

Second, I made slight changes to the magnet fractal types, that other
colours are used for points whose orbital converge to the finite
attractor at zero. To see these changes compare the file formulas.c.orig
with the file formulas.c.2 also found in the tar archive. The file
formulas.c.12 in the archive is used to build XaoS with both the first
and the second feature.

Third, I added the menuitem bailout just below the menuitem iterations
in the calculation menu. Its main purpose is to alter the escape radius
for infinite attractors. For an example compare the mandelbrot pictures
created with a bailout of 4 and 100 when the outcoloring mode smooth is
in use. Also the barnsley fractals show interesting variants with
various bailout values in effect. Bailout defaults to 4 and is limited
to values greater than or equal to zero. For magnet fractal types the
escape radius of the finite attractor at zero decreases with increasing
bailout values, to give the possibility to improve the precision around
the origin. To achieve this, I had to change the files
include/fconfig.h, include/fractal.h, include/ui_helper.h,
ui-hlp/menu.c, ui-hlp/save.c, ui-hlp/ui_helper.c, ui/ui.c,
ui/ui-drv/win32/xaos.dlg, engine/formulas.c and engine/fractal.c. All
their original and modified versions are included in the tar archive
with the extensions .orig and .3 respectively. The files formulas.c.13
and menu.c.13 in the archive are used to build XaoS with the first and
third feature and the file formulas.c.23 is used to build XaoS with the
second and the third feature. To build XaoS with all the features you
will use formulas.c.123.

Because some *.orig files are older than the current files in cvs, you
have to check if my changes conflict with changes that have been made
after I started to add my features.

With kind regards

Andreas