3 small patches by Philippe Wautelet <p.wautelet@fractalzone.be>, sent to me on Tue, 26 Oct 2004 22:44:12 +0200:
-----------------------8X-------------------------------------------------
Hi,

Here are the patches to correct some (small) bugs I found.

filesel.patch: correct a bug in the list of files in a directory: a
negative value could appear (the check was done too early) and was at
the origin of an access out of the valid memory range

ui.patch: correct a bug when the -nogui option was used and that there
are no available driver (this corrects the bug with the request ID 906942)

zoom.patch: permutated some conditions in 'if' to prevent access to not
initialized values

That's all for the moment.

Regards,

Philippe
-----------------------8X-------------------------------------------------

I compiled the code after applying the patches and no problems occured
during running (however I didn't do big tests, including the original
problems, too).